requirements.txt
ultralytics>=8.0.0
pillow>=10.0.0
opencv-python>=4.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
